requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.13.1